python/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
zstandard>=0.22.0
protobuf>=4.25.0
boto3>=1.34.0
